@@ -1,31 +1,32 @@
 name: veggie_android_seasons
 description: A new Flutter project.
 
-version: 2.0.0
+version: 3.0.0
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.19.6 <3.0.0"
 
 dependencies:
-  animations: ^2.0.1
-  cupertino_icons: ^1.0.3
+  animations: ^2.0.7
+  cupertino_icons: ^1.0.5
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.1.0
-  intl: ^0.17.0
-  provider: ^6.0.2
-  shared_preferences: ^2.0.5
+  font_awesome_flutter: ^10.4.0
+  intl: ^0.18.0
+  provider: ^6.0.5
+  shared_preferences: ^2.0.20
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.12.0
   flutter_lints: ^2.0.1
-  flutter_native_splash: ^2.1.1
+  flutter_native_splash: ^2.2.19
   flutter_test:
     sdk: flutter
 flutter_icons:
   android: true
   ios: true
   image_path: "android/app/src/main/res/drawable/launcher_icon.png"
+  remove_alpha_ios: true
 
 flutter_native_splash:
   color: "#FFFFFF"
@@ -34,7 +35,6 @@ flutter_native_splash:
   ios: false
   web: false
   android12: true
-
 
 # The following section is specific to Flutter.
 flutter:
